+++ SharpCompress/IO/SourceStream.cs
@@ -61,7 +61,7 @@ internal sealed class SourceStream : Stream
     // FenPhoenix 2023: avoid a zillion byte[1] allocations
     public override int ReadByte()
     {
-        return Read(FEN_COMMON.Byte1, 0, 1) == 0 ? -1 : (int)FEN_COMMON.Byte1[0];
+        return Read(FEN_COMMON.Byte1, 0, 1) == 0 ? -1 : FEN_COMMON.Byte1[0];
     }
 
     public override long Seek(long offset, SeekOrigin origin) => _stream.Seek(offset, origin);
